+++ .gitignore
@@ -3,6 +3,9 @@
 ##
 ## Get latest from https://github.com/github/gitignore/blob/master/VisualStudio.gitignore
 
+*utils
+
+
 # User-specific files
 *.rsuser
 *.suo
