--- utils.html
@@ -1,20 +1,0 @@
-<main>
-      <article class="grocery-wrapper">
-        <p class="alert"></p>
-        <header class="heading">
-          <h2 class="grocery-title">Today's Todo</h2>
-          <span><i class="fa-solid fa-rotate"></i></span>
-        </header>
-        <form class="grocery-form">
-          <input
-            type="text"
-            class="grocery-item"
-            placeholder="Add to your list"
-          />
-        </form>
-        <div class="grocery-container">
-          <div class="grocery-list"></div>
-          <button class="clear-list">clear items</button>
-        </div>
-      </article>
-    </main>
